Renaming UserProfileModal to Modal
