+++ src/components/AppSidebar/AppSidebar.tsx
@@ -13,7 +13,7 @@ import { getUserById } from '../../../firebase/functions/getUserById'
 import { getGoalBuddyByUserId } from '../../../firebase/functions/getGoalBuddyByUserId'
 import { GoalBuddy, User } from '@/types/types'
 import './AppSidebar.css'
-import UserProfileModal from '../UserProfileModal/UserProfileModal'
+import UserProfileModal from '../Modal/UserProfileModal'
 import { IdContext } from '../context/IdContext'
 import clsx from 'clsx'
 import { SidebarContext } from '../context/SidebarContext'
